+++ profiles-restclient/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Version History
 
+### 1.5
+
+ - Support added to look up property availability for a tag ID
+ - Added function to get the account realm from the service
+
 ### 1.4
 
  - Enable import of new tags by using the transaction handler