+++ utility/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Version History
 
+### 1.5
+
+ - Documentation updated
+
 ### 1.4
 
  - Added RNG in Secure.getRandomBytes(length)